udata: Move shared Utreexo data types and functions to a separate package
This resolves cycle imports that happen when importing Utreexo
to outside packages
