+++ accumulator/pollard.go
@@ -14,6 +14,30 @@ var ErrorStrings = map[uint32]error{
 	ErrorNoPollardNode:  fmt.Errorf("ErrorNoPollardNode"),
 }
 
+// Pollard is the sparse representation of the utreexo forest, using
+// binary tree pointers instead of a hash map.
+
+// I generally avoid recursion as much as I can, using regular for loops and
+// ranges instead.  That might start looking pretty contrived here, but
+// I'm still going to try it.
+
+// Pollard :
+type Pollard struct {
+	numLeaves uint64 // number of leaves in the pollard forest
+
+	roots []*polNode // slice of the tree roots, which are polNodes.
+	// roots are in big to small order
+	// BUT THEY'RE WEIRD!  The left / right children are actual children,
+	// not nieces as they are in every lower level.
+
+	hashesEver, rememberEver, overWire uint64
+
+	Lookahead int32 // remember leafs below this TTL
+	//	Minleaves uint64 // remember everything below this leaf count
+
+	positionMap map[MiniHash]uint64
+}
+
 // Modify is the main function that deletes then adds elements to the accumulator
 func (p *Pollard) Modify(adds []Leaf, delsUn []uint64) error {
 
@@ -455,6 +479,18 @@ func (p *Pollard) toFull() (*Forest, error) {
 	}
 
 	return ff, nil
+}
+
+// GetRoots returns the hashes of the pollard roots
+func (p *Pollard) GetRoots() (h []Hash) {
+	// pre-allocate. Shouldn't matter too much because this is only to export the
+	// utreexo state
+	h = make([]Hash, len(p.roots))
+
+	for _, pn := range p.roots {
+		h = append(h, pn.data)
+	}
+	return
 }
 
 func (p *Pollard) ToString() string {
