+++ csn/init.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/btcsuite/btcd/wire"
 	"github.com/mit-dci/utreexo/accumulator"
+	"github.com/mit-dci/utreexo/btcacc"
 	"github.com/mit-dci/utreexo/util"
 )
 
@@ -117,7 +118,7 @@ func (c *Csn) Start(cfg *Config, height int32, path, proxyURL string, haltSig ch
 // initCSNState attempts to load and initialize the CSN state from the disk.
 // If a CSN state is not present, chain is initialized to the genesis
 func initCSNState() (
-	p accumulator.Pollard, height int32, utxos map[wire.OutPoint]util.LeafData, err error) {
+	p accumulator.Pollard, height int32, utxos map[wire.OutPoint]btcacc.LeafData, err error) {
 
 	// bool to check if the pollarddata is present
 	pollardInitialized := util.HasAccess(PollardFilePath)
@@ -132,7 +133,7 @@ func initCSNState() (
 		fmt.Println("Creating new pollarddata")
 		// start at height 1
 		height = 1
-		utxos = make(map[wire.OutPoint]util.LeafData)
+		utxos = make(map[wire.OutPoint]btcacc.LeafData)
 		// Create file needed for pollard
 		_, err = os.OpenFile(PollardFilePath, os.O_CREATE, 0600)
 		if err != nil {
