+++ wire/doc.go
@@ -1,0 +1,3 @@
+// wire implements the network logic for transactions that include the
+// utreexo accumulator proofs
+package wire
